kernel + comp00 + demo00: The kernel is not loading icons anymore.

+++ np/heavy/games3d/demo00/libui/bmp.c
@@ -632,7 +632,7 @@ gwssrv_display_system_icon (
     sm_buffer = (char *) __get_system_icon(index);
     if ( (void *) sm_buffer == NULL ){
         printf ("gwssrv_display_system_icon: sm_buffer\n");
-        return -1;
+        return (int) -1;
     }
 
 // Check BM header
